tests/requirements.txt
openai>=1.50.0
telethon==1.34.0
cryptg==0.4.0
pytest-mock>=3.12.0

